+++ .env.qa
@@ -6,9 +6,23 @@ MONGODB_USERNAME=
 MONGODB_PASS=
 MONGODB_CLUSTER=
 NOTION_TOKEN=
-DISCORD_ROLE_GUEST_PASS=857683488131907624
-DISCORD_ROLE_LEVEL_2=858064540393275413
 FAQS_PAGE_ID=f99e764c-9eb7-4d87-bfc7-fcd93fa93df5
-NOTION_GUEST_PASS_DATABASE_ID=30087b23eab34e93a200ed3603eb5e95
 DISCORD_SERVER_ID=851552281249972254
+NOTION_GUEST_PASS_DATABASE_ID=30087b23eab34e93a200ed3603eb5e95
+
+# Roles
+DISCORD_ROLE_GENESIS_SQUAD=872217792626163804
+DISCORD_ROLE_ADMIN_ROLE_ID=870771734352113724
+DISCORD_ROLE_LVL_4=871597401687871508
+DISCORD_ROLE_LEVEL_3=870013262421450842
+DISCORD_ROLE_CONTRIBUTORS_LVL_2=858064540393275413
+DISCORD_ROLE_LEVEL_1=870012820207587368
+DISCORD_ROLE_GRANTS_COMITTEE=872217606961115177
+DISCORD_ROLE_DEVELOPERS_GUILD=857838853326241814
+DISCORD_ROLE_GUEST_PASS=857683488131907624
+
+# Channels
+DISCORD_CHANNEL_BOUNTY_BOARD_ID=869964294865977425
+
+# MISC
 DAO_CURRENT_SEASON=1
