+++ .github/workflows/deploy-prod.yml
@@ -43,13 +43,14 @@ jobs:
                   HD_FAQS_PAGE_ID: '29353e63-e864-411a-abc0-061d04906c16'
                   HD_DISCORD_SERVER_ID: '834499078434979890'
                   HD_NOTION_GUEST_PASS_DATABASE_ID: '15fae33760724cf7b696a981d53e50cb'
-                  HD_DISCORD_ROLE_GUEST_PASS: '848369657281773599'
-                  HD_DISCORD_ROLE_LEVEL_2: '839005084016312360'
                   HD_DAO_CURRENT_SEASON: '1'
-                  HD_DISCORD_ROLE_GRANTS_COMITTEE: '860222982416498688'
+                  HD_DISCORD_CHANNEL_BOUNTY_BOARD_ID: '850402063741091880'
                   HD_DISCORD_ROLE_GENESIS_SQUAD: '839004775751352341'
+                  HD_DISCORD_ROLE_ADMIN_ROLE_ID: '836317754038026241'
+                  HD_DISCORD_ROLE_GRANTS_COMITTEE: '860222982416498688'
                   HD_DISCORD_ROLE_LVL_4: '840627393777762344'
                   HD_DISCORD_ROLE_LEVEL_3: '839005209611075614'
                   HD_DISCORD_ROLE_CONTRIBUTORS_LVL_2: '839005084016312360'
                   HD_DISCORD_ROLE_LEVEL_1: '834560235704025108'
+                  HD_DISCORD_ROLE_GUEST_PASS: '848369657281773599'
                   HD_DISCORD_ROLE_DEVELOPERS_GUILD: '843924063591727195'
