+++ src/app/commands/admin/GuestPass.ts
@@ -1,6 +1,6 @@
 import { SlashCommand, CommandOptionType, ApplicationCommandPermissionType, CommandContext } from 'slash-create';
 import client from '../../app';
-import roleIDs from '../../constants/roleIDs';
+import roleIDs from '../../service/constants/roleIDs';
 import AddGuestPass, { expiresInHours } from '../../service/guest-pass/AddGuestPass';
 
 module.exports = class GuestPass extends SlashCommand {
