+++ src/app/service/bounty/DeleteBounty.ts
@@ -1,4 +1,4 @@
-import constants from '../../constants';
+import constants from '../constants/constants';
 import mongo, { Db, UpdateWriteOpResult } from 'mongodb';
 import ServiceUtils from '../../utils/ServiceUtils';
 import dbInstance from '../../utils/db';
