+++ src/app/service/guest-pass/AddGuestPass.ts
@@ -1,7 +1,7 @@
 import client from '../../app';
 import { Db } from 'mongodb';
 import dbInstance from '../../utils/db';
-import constants from '../../constants';
+import constants from '../constants/constants';
 import ServiceUtils from '../../utils/ServiceUtils';
 import { GuildMember } from 'discord.js';
